+++ src/backend/cdb/cdbmirroredappendonly.c
@@ -28,6 +28,7 @@
 #include "cdb/cdbpersistenttablespace.h"
 #include "cdb/cdbfilerepprimary.h"
 #include "storage/smgr.h"
+#include "storage/smgr_ao.h"
 #include "storage/lwlock.h"
 #include "utils/guc.h"
 #include "cdb/cdbpersistentfilesysobj.h"
